+++ arcane/src/arcane/alina/RichardsonSolver.h
@@ -45,7 +45,7 @@ struct RichardsonSolverParams
   double damping = 1.0;
 
   /// Maximum number of iterations.
-  size_t maxiter = 100;
+  Int32 maxiter = 100;
 
   /// Target relative residual error.
   double tol = 1.0e-8;
